Move video embed to subcomponent
